fix: resolve clippy warnings in source code
- Remove unnecessary cast in monitoring/persistence.rs
- Remove trim() before split_whitespace() in platform/linux.rs
- Fix needless borrows in platform/linux.rs and security/process_isolation.rs
- Fix unnecessary cast in security/resource_limits.rs
- Remove redundant import in security/permissions.rs
- Add Default implementation for ControlFlowIntegrity

+++ src/monitoring/persistence.rs
@@ -458,7 +458,7 @@ impl PersistenceManager {
             info!("Cleaned up {} old health check records", result);
         }
 
-        Ok(result as usize)
+        Ok(result)
     }
 }
